refactoring & cleaning up
