+++ inevolin/Config.sample.py
@@ -1,0 +1,12 @@
+config = {
+    
+    'KEY': 'change_me',
+    'SECRET': 'change_me',
+    'CONTAINER_NAME': 'change_me',
+
+    'MULTIPART_SIZE_MB': 500,
+    'MAX_THREADS': 200,
+    'DUMMY_SIZE_MB': 5,
+    'DUMMY_ALGO_MARGIN': 0.9,
+    'THREAD_TIMEOUT': None,
+}
